Add IKS v1.11+ support and add a deploy script
This commit adds support for IKS v1.11+. This will use a docker
deployment in Kubernetes since `docker` would no longer be used as the
container runtime of the worker nodes (docker.sock is no longer present)

This commit also adds in an easy to build and deploy script that would
help in quickly executing the commands for building and pushing Docker
images and deploying them to Kubernetes. This will make the steps for
the pattern shorter.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>